+++ lawmon/src/pages/chat/RoomItem.css
@@ -1,0 +1,15 @@
+.RoomContainer {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 10px;
+    border-bottom: 1px solid #ccc;
+    width: 1000px;
+}
+.chattingID {
+    font-size: 20px;
+}
+
+.lastMessage {
+    font-size: 15px;
+}
